# apps/api/swagger.json
{
  "openapi": "3.1.0",
  "paths": {
    "/organizations/{organizationId}/users/by-id": {
      "get": {
        "description": "Retrieve a user by their unique identifier within an organization.",
        "operationId": "GetUserByIdController",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          },
          {
            "name": "targetUserId",
            "required": false,
            "in": "query",
            "description": "The unique identifier of the user to retrieve",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with user details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GetUserByIdResponseDto_Output" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get user by id",
        "tags": ["Users"]
      }
    },
...
    "/organizations/{organizationId}/services": {
      "post": {
        "description": "Create a new service within an organization.",
        "operationId": "CreateServiceController",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateServiceBodyDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response with created service details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateServiceResponseDto_Output" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorDto" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create a new service",
        "tags": ["Services"]
      },
      "get": {
        "description": "Retrieve all services within an organization.",
        "operationId": "GetAllServicesController",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": { "type": "number" }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "Number of items per page for pagination",
            "schema": { "type": "number" }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Order of the results (asc or desc)",
            "schema": { "enum": ["asc", "desc"], "type": "string" }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "Field to order the results by",
            "schema": { "enum": ["name", "createdAt", "updatedAt"], "type": "string" }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter services by name",
            "schema": { "type": "string" }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with all services details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GetAllServicesResponseDto_Output" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorDto" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorDto" } }
            }
          }
        },
        "summary": "Get all services",
        "tags": ["Services"]
      }
    },
    "/organizations/{organizationId}/services/{serviceId}": {
      "delete": {
        "description": "Delete a service within an organization.",
        "operationId": "DeleteServiceController",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the service to be deleted",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with deleted service details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteServiceResponseDto_Output" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorDto" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Delete a service",
        "tags": ["Services"]
      },
      "put": {
        "description": "Update the details of a service within an organization.",
        "operationId": "UpdateServiceController",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the service",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateServiceBodyDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response with updated service details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UpdateServiceResponseDto_Output" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorDto" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update a service",
        "tags": ["Services"]
      }
    },
...
    "/organizations/{organizationId}/services/by-id/{serviceId}": {
      "get": {
        "description": "Retrieve a specific service within an organization by its ID.",
        "operationId": "GetServiceByIdController",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the service",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with the service details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GetServiceByIdResponseDto_Output" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorDto" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorDto" } }
            }
          }
        },
        "summary": "Get a service by ID",
        "tags": ["Services"]
      }
    },
    "/organizations/{organizationId}/tickets": {
      "post": {
        "description": "Create a new ticket within an organization.",
        "operationId": "CreateTicketController",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateTicketBodyDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response with created ticket details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateTicketResponseDto_Output" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorDto" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorDto" } }
            }
          }
        },
        "summary": "Create a new ticket",
        "tags": ["Tickets"]
      }
    },
...
    "/organizations/{organizationId}/tickets/{ticketId}/transfer": {
      "put": {
        "description": "Transfer a ticket to another service within an organization. Optionally specify a target staff member.",
        "operationId": "TransferTicketController",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          },
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the ticket",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TransferTicketBodyDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response with transferred ticket details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TransferTicketResponseDto_Output" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorDto" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Transfer a ticket to another service or colleague",
        "tags": ["Tickets"]
      }
    },
...
  },
  "info": {
    "title": "Queue management",
    "description": "A queue management API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
      "NotFoundErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "const": 404 },
          "message": { "type": "string" }
        },
        "required": ["statusCode", "message"]
      },
      "GetUserByIdResponseDto_Output": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "ulid",
                "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
              },
              "name": { "type": "string" },
              "email": { "type": "string" },
              "emailVerified": { "type": "boolean" },
              "banned": { "type": "boolean" },
              "banReason": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
              "banExpires": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
              "image": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
              "role": { "type": "string" },
              "createdAt": { "type": "string" },
              "updatedAt": { "type": "string" }
            },
            "required": [
              "id",
              "name",
              "email",
              "emailVerified",
              "banned",
              "banReason",
              "banExpires",
              "image",
              "createdAt",
              "updatedAt"
            ],
            "additionalProperties": false
          }
        },
        "required": ["user"],
        "additionalProperties": false
      },
      "UnauthorizedErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "const": 401 },
          "message": { "type": "string" }
        },
        "required": ["statusCode", "message"]
      },
...
      "CreateServiceBodyDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 255 },
          "description": { "type": "string" },
          "avgDurationInt": { "type": "number" },
          "isActive": { "type": "boolean" },
          "maxCapacity": { "type": "number" }
        },
        "required": ["name"]
      },
      "CreateServiceResponseDto_Output": {
        "type": "object",
        "properties": {
          "service": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "ulid",
                "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
              },
              "name": { "type": "string" },
              "description": { "type": "string" },
              "avgDurationInt": { "type": "number" },
              "maxCapacity": { "anyOf": [{ "type": "number" }, { "type": "null" }] },
              "alertThresholdMinutes": { "type": "number" },
              "isActive": { "type": "boolean" },
              "organizationId": { "type": "string" },
              "createdAt": { "type": "string" },
              "updatedAt": { "type": "string" }
            },
            "required": ["id", "name", "organizationId", "createdAt", "updatedAt"],
            "additionalProperties": false
          }
        },
        "required": ["service"],
        "additionalProperties": false
      },
      "DeleteServiceResponseDto_Output": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "required": ["message"],
        "additionalProperties": false
      },
      "GetAllServicesResponseDto_Output": {
        "type": "object",
        "properties": {
          "services": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "ulid",
                  "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
                },
                "name": { "type": "string" },
                "description": { "type": "string" },
                "avgDurationInt": { "type": "number" },
                "maxCapacity": { "anyOf": [{ "type": "number" }, { "type": "null" }] },
                "alertThresholdMinutes": { "type": "number" },
                "isActive": { "type": "boolean" },
                "organizationId": { "type": "string" },
                "createdAt": { "type": "string" },
                "updatedAt": { "type": "string" }
              },
              "required": ["id", "name", "organizationId", "createdAt", "updatedAt"],
              "additionalProperties": false
            }
          },
          "total": { "type": "number" }
        },
        "required": ["services", "total"],
        "additionalProperties": false
      },
...
      "UpdateServiceBodyDto": {
        "type": "object",
        "properties": {
          "isActive": { "type": "boolean" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "avgDurationInt": { "type": "number" }
        }
      },
      "UpdateServiceResponseDto_Output": {
        "type": "object",
        "properties": {
          "service": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "ulid",
                "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
              },
              "name": { "type": "string" },
              "description": { "type": "string" },
              "avgDurationInt": { "type": "number" },
              "maxCapacity": { "anyOf": [{ "type": "number" }, { "type": "null" }] },
              "alertThresholdMinutes": { "type": "number" },
              "isActive": { "type": "boolean" },
              "organizationId": { "type": "string" },
              "createdAt": { "type": "string" },
              "updatedAt": { "type": "string" }
            },
            "required": ["id", "name", "organizationId", "createdAt", "updatedAt"],
            "additionalProperties": false
          }
        },
        "required": ["service"],
        "additionalProperties": false
      },
      "GetServiceByIdResponseDto_Output": {
        "type": "object",
        "properties": {
          "service": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "ulid",
                "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
              },
              "name": { "type": "string" },
              "description": { "type": "string" },
              "avgDurationInt": { "type": "number" },
              "maxCapacity": { "anyOf": [{ "type": "number" }, { "type": "null" }] },
              "alertThresholdMinutes": { "type": "number" },
              "isActive": { "type": "boolean" },
              "organizationId": { "type": "string" },
              "createdAt": { "type": "string" },
              "updatedAt": { "type": "string" }
            },
            "required": ["id", "name", "organizationId", "createdAt", "updatedAt"],
            "additionalProperties": false
          }
        },
        "required": ["service"],
        "additionalProperties": false
      },
      "CreateTicketBodyDto": {
        "type": "object",
        "properties": {
          "guestName": { "type": "string", "minLength": 1, "maxLength": 255 },
          "serviceId": {
            "type": "string",
            "format": "ulid",
            "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
          }
        },
        "required": ["guestName", "serviceId"]
      },
      "CreateTicketResponseDto_Output": {
        "type": "object",
        "properties": {
          "ticket": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "ulid",
                "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
              },
              "guestName": { "type": "string" },
              "status": {
                "type": "string",
                "enum": ["WAITING", "CALLED", "SERVING", "SERVED", "ABSENT", "CANCELLED"]
              },
              "callCount": { "type": "number" },
              "organizationId": { "type": "string" },
              "serviceId": { "type": "string" },
              "servedById": { "type": "string" },
              "joinedAt": { "type": "string" },
              "calledAt": { "type": "string" },
              "startedAt": { "type": "string" },
              "completedAt": { "type": "string" }
            },
            "required": ["id", "guestName", "status", "organizationId", "serviceId", "joinedAt"],
            "additionalProperties": false
          }
        },
        "required": ["ticket"],
        "additionalProperties": false
      },
...
      "TransferTicketBodyDto": {
        "type": "object",
        "properties": {
          "targetServiceId": {
            "type": "string",
            "format": "ulid",
            "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
          },
          "targetStaffId": {
            "type": "string",
            "format": "ulid",
            "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
          }
        },
        "required": ["targetServiceId"]
      },
      "TransferTicketResponseDto_Output": {
        "type": "object",
        "properties": {
          "ticket": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "ulid",
                "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
              },
              "guestName": { "type": "string" },
              "status": {
                "type": "string",
                "enum": ["WAITING", "CALLED", "SERVING", "SERVED", "ABSENT", "CANCELLED"]
              },
              "callCount": { "type": "number" },
              "organizationId": { "type": "string" },
              "serviceId": { "type": "string" },
              "servedById": { "type": "string" },
              "joinedAt": { "type": "string" },
              "calledAt": { "type": "string" },
              "startedAt": { "type": "string" },
              "completedAt": { "type": "string" }
            },
            "required": ["id", "guestName", "status", "organizationId", "serviceId", "joinedAt"],
            "additionalProperties": false
          }
        },
        "required": ["ticket"],
        "additionalProperties": false
      },
...
    }
  }
}